% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc sampleing.R
\name{mcmcsample}
\alias{mcmcsample}
\title{MCMC sample from a pdf}
\usage{
mcmcsample(t, pdf)
}
\arguments{
\item{t}{a sample of numbers}

\item{pdf}{a random variable}

\item{n}{a number}
}
\value{
the samples
}
\description{
MCMC sample from a pdf
}
\examples{
pdf<-function(x){x/2}
t<-runif(1,0,2)
mcmcsample(t, pdf)
}
