% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rejsample}
\alias{rejsample}
\title{Rejection sample from a pdf}
\usage{
rejsample(n, pdf, a, b, C)
}
\arguments{
\item{n}{a number}

\item{pdf}{a random variable}

\item{a}{a number}

\item{b}{a number}

\item{C}{a number}
}
\value{
the samples
}
\description{
Rejection sample from a pdf
}
\examples{
pdf<-function(x){x/2}
rejsample(2, pdf, 0, 5, .5)
}
